adicionado servico de taxa

+++ api-multbanc/src/main/java/com/multBancapp/apimultbanc/entities/AccountEntity.java
@@ -38,7 +38,7 @@ public class AccountEntity {
 
       private  BigDecimal performace;
 
-      private Double rate;
+      private BigDecimal rate;
 
 
       public void deposit(BigDecimal value) {
